Upgraded dojo to 1.6, cleaned out docs, changed code to use AMD module loading like dojo does, started work on programming the global toolbar widget

+++ dojotoolkit/dijit/PopupMenuItem.js
@@ -1,18 +1,16 @@
-dojo.provide("dijit.PopupMenuItem");
-
-dojo.require("dijit.MenuItem");
+define("dijit/PopupMenuItem", ["dojo", "dijit", "dijit/MenuItem"], function(dojo, dijit) {
 
 dojo.declare("dijit.PopupMenuItem",
 		dijit.MenuItem,
 		{
 		_fillContent: function(){
-			// summary: 
+			// summary:
 			//		When Menu is declared in markup, this code gets the menu label and
 			//		the popup widget from the srcNodeRef.
 			// description:
 			//		srcNodeRefinnerHTML contains both the menu item text and a popup widget
 			//		The first part holds the menu item text and the second part is the popup
-			// example: 
+			// example:
 			// |	<div dojoType="dijit.PopupMenuItem">
 			// |		<span>pick me</span>
 			// |		<popup> ... </popup>
@@ -40,6 +38,7 @@ dojo.declare("dijit.PopupMenuItem",
 				this.popup = dijit.byNode(node);
 			}
 			dojo.body().appendChild(this.popup.domNode);
+			this.popup.startup();
 
 			this.popup.domNode.style.display="none";
 			if(this.arrowWrapper){
@@ -47,13 +46,20 @@ dojo.declare("dijit.PopupMenuItem",
 			}
 			dijit.setWaiState(this.focusNode, "haspopup", "true");
 		},
-		
+
 		destroyDescendants: function(){
 			if(this.popup){
-				this.popup.destroyRecursive();
+				// Destroy the popup, unless it's already been destroyed.  This can happen because
+				// the popup is a direct child of <body> even though it's logically my child.
+				if(!this.popup._destroyed){
+					this.popup.destroyRecursive();
+				}
 				delete this.popup;
 			}
 			this.inherited(arguments);
 		}
 	});
 
+
+return dijit.PopupMenuItem;
+});
